module-smsnotify: Set priority to 999 for receiver
Set the priority for the SMS_RECEIVED broadcast receiver to 999. The old
XML date was incorrect, as the priority attribute was part of the action
element instead of the intent-filter element. And 1010 was an incorrect
value anyways, since the maximum is 999 as per the updated documentation
[1].

Also SMS_RECEIVED broadcasts are non-abortable, at least as of Android
4.4 [2]. This means that it shouldn't matter at which priority the
receiver runs. But first tests with SMS enabled Hangouts 2.0 showed that
the Receiver was only run with a priority of 999 and 500, but not 1.

Footnotes:
[1]  http://developer.android.com/guide/topics/manifest/intent-filter-element.html

[2]  http://developer.android.com/about/versions/android-4.4.html#SMS