Define used permissions on the Manifests
Every custom permission should also be defined in the Manifest that
declares it as 'uses-permission' to avoid that particular permission not
being granted. For example, as user that installs smssend without having
contactsread installed will not be able to have smssend use contactsread.
The reason is that the o.p.p.READ_CONTACTS perm was not granted to
smssend, because the system was not aware of that permission as we only
defined it contactsread, which was not installed at that time.