transport-xmpp: Save password on Activity unfocus
Many users currently run into the pitfall that they enter their JID, the
devices JID and the devices password and hit the back button, while the
password EditText is still focused. This means that EditTextWatcher's
onFocusChange() did not run and the password has not been saved to the
SharedPreferences.

We know save whatever currently is entered as password when the Activity
loses focus by using the onPause() method.

Note that the JID EditFields are still affected by the problem but
- we have to verify the JIDs for correctness
- and they appear visually before the password EditText
so this approach is not feasible for the JID EditFields.