Add "Permission Check" feature and status GUI element
This moves the declaration of MAXS Permissions from the non-main
components to a single place: the main component.

Then we add a simple status GUI element, that informs the user about
problems with regard to permissions used by MAXS. A friendly smiley will
inform the user that everything is OK. A sad one hopefully motivates him
to fix the problems, which can be easily done by clicking on it, which
will trigger a Dialog to appear with click-able links to the Application
Settings activity of the offending application (which should be
removed and re-installed in most cases).