transport-xmpp: Use non-constant reconnect delay
I'm not 100% how the reconnect calculation is done with this this
change. At first it's a constant delay of 10 seconds for the first 60
retries (=10min). Then we use basically

10 seconds * (<reconnectionAttemptCountAfterConstant>^1.2)

and cap the maximum delay by 30 minutes.

Ideal we would use a function that starts with small increases, but
grows larger after some time.