transport-xmpp: Use Handler.hasMessages(int)
to determine if an action is really necessary.
For example a STOP_SERVICE action followed by a START_SERVICE action
will be now ignored.

Also changed NetworkConnectivityReceiver to send three different types
of actions. All those are now ignored by TransportService if there is
another action of the same type queued.

On NETWORK_TYPE_CHANGED, we also test now if the connection is really
down, by sending a ping with a 500ms timeout. It could be that the
connection got already re-established by the ConnectionClosedOnError()
invocation. This should prevent unnecessary re-connects.