MAXSTranspoertService: Restarte state
It appears that sticky services are not always restarted with a 'null'
intent. Sometimes they are restarted with an intent that is already
pending. In which case we would not restore the service state.

Now MAXSTransportService uses SharedPreferences to save the service
state between forced service restarts by the system. Differences from
the actual state are detected and corrected (e.g. START_SERVICE is
issued).