transport-xmpp: Register receiver on demand
Upcoming changes in Android N (likely API 24) will prevent
BroadcastReceivers listening for CONNECTIVITY_ACTION from being called
if they are registered in the Manifest (and if the app targets the API
level). This change prepares for this behavior change, and also makes
MAXS Transport XMPP more resource friendly on older Android devices in
case the Transport is not running (active).