Ignore Parcel unmarshalling errors in certain situations
Android 13 apparently changed the Parcel format. This is the first
time where I witness such a change. Since MAXS uses Parcel(ables) to
serialize objects into bytes for storage in databases, the
deserialization/unmarshalling of those objects fails if they are
stored in an older format. To be clear: Android's documentation
clearly states that Parcel(able)s are only meant for inter-process
communication, so this breaking MAXS is 100% MAXS' fault.

I guess it was just to tempting 10 years ago, when I wrote this, to
use Parcel for this case. And, it is also not fatal if we are not able
to recover the data. Hence this is what this commit does. We simply
ignore the cases where we can not unmarshal the Parcels from the
bytes at rest.