Migrate to central build-logic, set Android target compatiblity to 24, bump Smack
Migrating to central build-logic removes a lot of duplicate (build)
code.

This also tries to prepare for a higher target SDK. However, MAXS'
design consiting of modular background services does not seem to fit
well with modern Android concepts. I have sank much time trying to get
it work, but a target SDK of 24 is the best I could achieve so far.